tools/requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord==1.0.1
discord.py==1.6.0
gitdb==4.0.5
GitPython==3.1.14
idna==3.1
multidict==5.1.0
smmap==3.0.5
typing-extensions==3.7.4.3
yarl==1.6.3
